requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
deep-translator==1.7.0
dnspython==2.2.0
google-trans-new==1.1.9
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
importlib-metadata==4.11.1
libretranslatepy==2.1.1
lxml==4.8.0
pip==21.2.4
pycparser==2.21
pymongo==3.11.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
tabulate==0.8.9
tornado==6.1
translate==3.6.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.7.0
